requirements.txt
fastapi
uvicorn
python-multipart
Pillow
google-generativeai